commit by huangsc

+++ src/main/java/com/thinkgem/jeesite/modules/contract/service/LeaseContractService.java
@@ -144,7 +144,7 @@ public class LeaseContractService extends CrudService<LeaseContractDao, LeaseCon
 			List<LeaseContractDtl> list = leaseContractDtlDao.findList(leaseContractDtl);
 			for (LeaseContractDtl tmpLeaseContractDtl : list) {
 				// 计算开始日期与结束日期之间的月数
-				int month = DateUtils.getMonthSpace(tmpLeaseContractDtl.getStartDate(),
+				float month = DateUtils.getMonthSpace(tmpLeaseContractDtl.getStartDate(),
 						tmpLeaseContractDtl.getEndDate());
 				month = (month == 0 ? month++ : month);
 				for (int i = 1; i <= month; i++) {
